#393 feat: Move enums to enums directory
