+++ src/shared/constants/enums/filter-component.enum.ts
