+++ src/shared/constants/enums/index.ts
@@ -1,0 +1,2 @@
+export * from './enums';
+export * from './filter-component.enum';
