+++ src/shared/constants/index.ts
@@ -1,7 +1,6 @@
 export * from './api.constants';
 export * from './date.constants';
 export * from './enums';
-export * from './filter-component.enum';
 export * from './gotchi.constants';
 export * from './homepage.constants';
 export * from './numbers.constants';
